screen: Increase typical alt-tab icon size to 96
Now we removed thumbnailing in the alt-tab switcher, the switcher is
more difficult to use with its small icons.

We switch to mutter-like higher resolution icons to be able to distinct
them better (the new size is about the same order as the thumbnails),
as more and more applications are providing high-res icons.

This patch is ported from mutter, cf
https://bugzilla.gnome.org/show_bug.cgi?id=689651

https://bugzilla.gnome.org/show_bug.cgi?id=504729